requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
peewee==3.17.5
plumbum==1.8.3
python-telegram-bot==21.2
rpyc==6.0.0
sniffio==1.3.1
typing_extensions==4.11.0
yarl==1.9.4
